requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
Pillow
grpcio
numpy
opencv-python
scipy
tensorflow-serving-api
azure-storage-blob
